% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dendrogram.R
\name{as.ndendrogram}
\alias{as.ndendrogram}
\title{ndendrogram object converter}
\usage{
as.ndendrogram(d)
}
\arguments{
\item{d}{- Base or dendextend Dendrogram object from R}
}
\value{
ndendrogram object
}
\description{
utility to make a usable dendrogram for more complex plotting
}
