% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_annotate}
\alias{lp_annotate}
\title{Plot annotation layer}
\usage{
lp_annotate(
  lp,
  side,
  attribute,
  col = NULL,
  size = 1,
  gap = 0.4,
  palette = "Viridis",
  colors = NULL,
  show_bounding_box = F,
  type = "rect",
  label = attribute,
  label_just = "right",
  cex.label = 0.8,
  border = NA,
  group.border = NA,
  cex.point = 1,
  pch = 19,
  bg = NA,
  lwd = 1,
  show_label = T,
  zlim = NULL,
  reverse_palette = F,
  zero_centered_colors = F
)
}
\arguments{
\item{lp}{layermap object.}

\item{side}{value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{name for the attribute which will be plotted in the layer. For group, this must be defined in column_groups or row_groups.}

\item{col}{a named vector, where colors are the values and names are conditions that are defined in the attribute/group that is plotted.}

\item{size}{the space (in lines) for this layer in the margin.}

\item{gap}{the space (in lines) between this layer and the prior layer.}

\item{palette}{hcl.colors palette to fill in unnamed conditions colors.}

\item{show_bounding_box}{shows the bounding space for this layer (useful for troubleshooting)}

\item{cex.label}{- cex value for label characters.}

\item{layer_just}{- the justification side for a layer label (right or left).}
}
\value{
layermap object
}
\description{
Function for plotting an annotation layer based on column or row attributes.
}
