% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{lp_boundaries}
\alias{lp_boundaries}
\title{Layer boundaries}
\usage{
lp_boundaries(
  lp,
  side,
  size,
  gap,
  text.gap = 0,
  text.restriction = F,
  show_bounding_box = F
)
}
\description{
Utility for defining layer boundaries
}
