% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layermap.R
\name{lp_color_legend}
\alias{lp_color_legend}
\title{Plot color gradient legend}
\usage{
lp_color_legend(lp, side, size = 1, gap = 0.4, ratio = 3, adj = 0, round = 1)
}
\arguments{
\item{lp}{- layermap object}

\item{add}{- logical for whether this should plot a new window. Experimental.}
}
\description{
Makes a simple color gradient legend corresponding to the main heatmap.
}
