% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_dend}
\alias{lp_dend}
\title{Plot dendrogram layer}
\usage{
lp_dend(
  lp,
  side,
  size = 2,
  gap = 0.2,
  cutoff = T,
  cex = 0.8,
  show_bounding_box = F,
  ...
)
}
\arguments{
\item{lp}{layermap object.}

\item{side}{value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{size}{the space (in lines) for this layer in the margin.}

\item{gap}{the space (in lines) between this layer and the prior layer.}

\item{show_bounding_box}{shows the bounding space for this layer (useful for troubleshooting)}
}
\value{
layermap object
}
\description{
Function for plotting an dendrogram layer. This function will only work on sides which have been clustered by hclust in the initial layermap call.
}
