% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{lp_example_2}
\alias{lp_example_2}
\title{Test example 2}
\usage{
lp_example_2(save = F)
}
\description{
Makes a layermap based on the a layermap dataset of Nitrogen- and Drought-related genes in these treatments.
}
