% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{lp_example_3}
\alias{lp_example_3}
\title{Test example 3}
\usage{
lp_example_3()
}
\description{
Makes a layermap based on the a state database. This showcases multi-divisions and separated clustering compared to clustered grouping of axes.
}
