% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{lp_example_4}
\alias{lp_example_4}
\title{Test example 4}
\usage{
lp_example_4()
}
\description{
Makes a layermap based on the a fivethirtyeight airline_safety dataset. This is meant to show more examples of annotation layers.
}
