% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_group}
\alias{lp_group}
\title{Plot group layer}
\usage{
lp_group(
  lp,
  side,
  attribute,
  size = 1,
  gap = 0.4,
  palette = "Zissou 1",
  col = NULL,
  show_bounding_box = F,
  plot_names = T,
  cex.names = 0.8,
  plot_label = T,
  cex.label = 0.8,
  label_just = "right"
)
}
\arguments{
\item{lp}{layermap object.}

\item{side}{value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{name for the attribute which will be plotted in the layer. For group, this must be defined in column_groups or row_groups.}

\item{size}{the space (in lines) for this layer in the margin.}

\item{gap}{the space (in lines) between this layer and the prior layer.}

\item{palette}{hcl.colors palette to fill in unnamed conditions colors.}

\item{col}{a named vector, where colors are the values and names are conditions that are defined in the attribute/group that is plotted.}

\item{show_bounding_box}{shows the bounding space for this layer (useful for troubleshooting)}

\item{plot_names}{logical for whether names should be plotted for groups.}

\item{cex.names}{cex value relating to name text size.}

\item{plot_label}{logical for whether layer labels should be plotted.}

\item{cex.label}{cex value relating to layer label text size.}

\item{label_just}{the justification side for a layer label (right or left).}
}
\value{
layermap object
}
\description{
Function for plotting a group layer based on column or row attributes.
}
