% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_group_names}
\alias{lp_group_names}
\title{Plot group name layer}
\usage{
lp_group_names(
  lp,
  side,
  attribute,
  col = "black",
  font = 1,
  adj = NULL,
  family = "",
  srt = NULL,
  size = 1,
  gap = 0,
  cex = 0.8,
  show_bounding_box = F,
  group_label = F
)
}
\arguments{
\item{lp}{layermap object.}

\item{side}{value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{name for the attribute which will be plotted in the layer. For group, this must be defined in column_groups or row_groups.}

\item{col}{a named vector, where colors are the values and names are conditions that are defined in the attribute/group that is plotted.}

\item{srt}{}

\item{size}{the space (in lines) for this layer in the margin.}

\item{gap}{the space (in lines) between this layer and the prior layer.}

\item{cex}{- cex value for label characters.}

\item{show_bounding_box}{shows the bounding space for this layer (useful for troubleshooting)}

\item{palette}{hcl.colors palette to fill in unnamed conditions colors.}
}
\value{
layermap object
}
\description{
Function for plotting a name label for a group.
}
