% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_group_pie}
\alias{lp_group_pie}
\title{Plot group pie layer}
\usage{
lp_group_pie(
  lp,
  side,
  attribute,
  col = NULL,
  palette = "Zissou 1",
  size = 1,
  gap = 0.4,
  cex = 1,
  show_bounding_box = F,
  label_just = "right",
  group_label = T,
  cex.label = 0.8,
  trace = T,
  trace.lty = 1,
  trace.lwd = 0.8,
  trace.gap = 1.5,
  trace.col = "black"
)
}
\arguments{
\item{lp}{layermap object.}

\item{side}{value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{name for the attribute which will be plotted in the layer. For group, this must be defined in column_groups or row_groups.}

\item{col}{a named vector, where colors are the values and names are conditions that are defined in the attribute/group that is plotted.}

\item{palette}{hcl.colors palette to fill in unnamed conditions colors.}

\item{size}{the space (in lines) for this layer in the margin.}

\item{gap}{the space (in lines) between this layer and the prior layer.}

\item{show_bounding_box}{shows the bounding space for this layer (useful for troubleshooting)}

\item{cex_label}{- cex value for label characters.}

\item{labels}{- logical for whether group labels should be plotted.}
}
\value{
layermap object
}
\description{
Function for plotting a group as a pie chart based on column or row attributes.
}
