% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layermap.R
\name{lp_legend}
\alias{lp_legend}
\title{Plot simple categorical legend}
\usage{
lp_legend(lp, add = F)
}
\arguments{
\item{lp}{- layermap object}

\item{add}{- logical for whether this should plot a new window. Experimental.}
}
\description{
Makes a simple heatmap color legend for a layermap object.
}
