% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{lp_line_to_coord}
\alias{lp_line_to_coord}
\title{Convert plot lines to coordinates}
\usage{
lp_line_to_coord(xmax, ymax, side, line)
}
\value{
value in coordinates
}
\description{
Utility for converting plotting coordinate systems.
}
