% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_plot_values}
\alias{lp_plot_values}
\title{Plot values}
\usage{
lp_plot_values(
  lp,
  alt.df = NULL,
  round.n = 2,
  cex = 0.6,
  light_color = "white",
  dark_color = "black",
  l_threshold = 50
)
}
\arguments{
\item{lp}{layermap object.}

\item{alt.df}{an alternative data.frame which will provide the values. Useful if the figure is colored with a different value than you want to show inside (for example scaled data). Should be in the same format as the value.df. NAs are not plotted}

\item{round.n}{number of decimals included in value.}

\item{cex}{text size.}

\item{light_color}{text color for dark backgrounds.}

\item{dark_color}{text color for light backgrounds.}

\item{l_threshold}{lightness value in lab of background color to choose if values are plotted as the light color or dark color.}
}
\value{
layermap object
}
\description{
Function for plotting values over main heatmap
}
