% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{lr_color_legend}
\alias{lr_color_legend}
\title{Plot color gradient legend}
\usage{
lr_color_legend(
  lr,
  side,
  size = 1,
  gap = 0.4,
  text.gap = 1,
  ratio = 3,
  just = "lb"
)
}
\arguments{
\item{lr}{- layermap object}

\item{add}{- logical for whether this should plot a new window. Experimental.}
}
\description{
Makes a simple color gradient legend corresponding to the main heatmap.
}
