% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{lr_colorize}
\alias{lr_colorize}
\title{Build a named color vector}
\usage{
lr_colorize(col, conditions, palette)
}
\arguments{
\item{col}{- named color vector}

\item{conditions}{- list of conditions to colorize}

\item{palette}{- hcl.colors palette used to fill in unnamed colors}
}
\value{
named color vector
}
\description{
Utility for getting named color vector used in layermap layer functions.
}
