% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{lr_legend}
\alias{lr_legend}
\title{Plot simple legend}
\usage{
lr_legend(lr, add = F)
}
\arguments{
\item{lr}{- layermap object}

\item{add}{- logical for whether this should plot a new window. Experimental.}
}
\description{
Makes a simple heatmap color legend for a layermap object.
}
