% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{lr_line_to_coord}
\alias{lr_line_to_coord}
\title{Convert plot lines to coordinates}
\usage{
lr_line_to_coord(xmax, ymax, side, line)
}
\value{
value in coordinates
}
\description{
Utility for converting plotting coordinate systems.
}
