% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{ly_dend}
\alias{ly_dend}
\title{Plot dendrogram layer}
\usage{
ly_dend(
  ly,
  side,
  size = 1,
  gap = 0.2,
  cutoff = T,
  cex = 0.8,
  show_bounding_box = F,
  ...
)
}
\arguments{
\item{ly}{- layermap object.}

\item{side}{- value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{prop}{- the proportion of plotting space reserved for this layer}
}
\value{
layermap object
}
\description{
Function for plotting an dendrogram layer. This function will only work on sides which have been clustered by hclust in the initial layermap call.
}
