% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{ly_label}
\alias{ly_label}
\title{Plot label on annotate or group layers}
\usage{
ly_label(ly, x_vec, y_vec, side, text, just, offset = 0.9, cex)
}
\arguments{
\item{to}{be added}
}
\value{
named color vector
}
\description{
Utility for plotting labels on layers
}
