% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{ly_legend}
\alias{ly_legend}
\title{Plot simple legend}
\usage{
ly_legend(ly, add = F)
}
\arguments{
\item{ly}{- layermap object}

\item{add}{- logical for whether this should plot a new window. Experimental.}
}
\description{
Makes a simple heatmap color legend for a layermap object.
}
