% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND_data.R
\docType{data}
\name{ND}
\alias{ND}
\title{Test Nitrogen/Drought dataset}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(ND)
}
\description{
Data from a meta-analsis of N and D projects in the SRA. L2FC data for a set of genes and libraries in a matrix. Includes values, and metadata for genes and libraries.
}
\examples{
data(ND)

}
\keyword{datasets}
