% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nh_boundaries}
\alias{nh_boundaries}
\title{empty for now}
\usage{
nh_boundaries(
  nh,
  side,
  size,
  gap,
  text.gap = 0,
  text.restriction = F,
  show_bounding_box = F,
  din_adjusted = T
)
}
\arguments{
\item{still}{empty}
}
\description{
empty for now
}
