% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap}
\alias{nheatmap}
\title{empty for now}
\usage{
nheatmap(
  value.df,
  xlim = NULL,
  ylim = NULL,
  column.df = NULL,
  row.df = NULL,
  column_groups = c(),
  row_groups = c(),
  palette = "PuOr",
  reverse_palette = T,
  zero_centered_colors = F,
  cluster_cols = F,
  cluster_rows = T,
  group_gap = 0.02,
  border = "grey25",
  plot_margin = c(0.2, 0.2, 0.2, 0.2),
  dim_reference = "din"
)
}
\arguments{
\item{still}{empty}
}
\description{
empty for now
}
