% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap_annotate}
\alias{nheatmap_annotate}
\title{empty for now}
\usage{
nheatmap_annotate(
  nh,
  side,
  aname,
  a.df = NULL,
  col = NULL,
  prop = 0.05,
  palette = "Viridis",
  show_bounding_box = F,
  type = "rect",
  label_just = "right",
  cex.label = 0.8
)
}
\arguments{
\item{still}{empty}
}
\description{
empty for now
}
