% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap_boundaries}
\alias{nheatmap_boundaries}
\title{empty for now}
\usage{
nheatmap_boundaries(nh, side, prop, show_bounding_box = F, din_adjusted = T)
}
\arguments{
\item{still}{empty}
}
\description{
empty for now
}
