% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap_dend}
\alias{nheatmap_dend}
\title{empty for now}
\usage{
nheatmap_dend(
  nh,
  side,
  prop = 0.1,
  cutoff = T,
  cex = 0.8,
  show_bounding_box = F,
  ...
)
}
\arguments{
\item{still}{empty}
}
\description{
empty for now
}
