% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap_group}
\alias{nheatmap_group}
\title{empty for now}
\usage{
nheatmap_group(
  nh,
  side,
  gname,
  col = NULL,
  palette = "Zissou 1",
  prop = NULL,
  cex = 0.8,
  show_bounding_box = F,
  label_just = "right",
  labels = T,
  cex.label = 0.8
)
}
\arguments{
\item{still}{empty}
}
\description{
empty for now
}
