% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap_legend}
\alias{nheatmap_legend}
\title{Plot simple legend}
\usage{
nheatmap_legend(nh, add = F)
}
\arguments{
\item{nh}{- nheatmap object}

\item{add}{- logical for whether this should plot a new window. Experimental.}
}
\description{
Makes a simple heatmap color legend for a nheatmap object.
}
