% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap_names}
\alias{nheatmap_names}
\title{empty for now}
\usage{
nheatmap_names(
  nh,
  side,
  aname = F,
  names = NULL,
  prop = 0.1,
  cutoff = T,
  cex = 0.8,
  show_bounding_box = F
)
}
\arguments{
\item{still}{empty}
}
\description{
empty for now
}
