% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layermap.R
\name{plot.ndendrogram}
\alias{plot.ndendrogram}
\title{ndendrogram plotting function}
\usage{
\method{plot}{ndendrogram}(
  d,
  horiz = T,
  flip.x = F,
  flip.y = F,
  type = "square",
  add = F,
  scale.x = NULL,
  scale.y = NULL,
  pos.x = 0,
  pos.y = 0,
  ...
)
}
\arguments{
\item{d}{- an ndendrogram object}

\item{horiz}{- logical assigns whether the dendrogram is plotted vertical or horizontal.}
}
\description{
plot.ndendrogram plots an ndendgrogram object in specific locations
}
