% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{vector_to_colors}
\alias{vector_to_colors}
\title{Utility to produce a color palette}
\usage{
vector_to_colors(
  values,
  zlim = NULL,
  na.color = "grey",
  zero_centered_colors = F,
  palette = "viridis",
  reverse_palette = F
)
}
\value{
a vector of colors
}
\description{
A simple utility used by many lp functions to define color scales for a vector using an hcl.palette.
}
