% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_color_legend}
\alias{lp_color_legend}
\title{Plot color gradient legend}
\usage{
lp_color_legend(
  lp,
  side,
  attributes = c("main"),
  size = 0.5,
  gap = 1.5,
  size_p = 0.4,
  gap_p = 0.05,
  round = 1,
  cex = 0.6,
  cex.title = NULL,
  titles = attributes
)
}
\arguments{
\item{lp}{- layermap object}

\item{side}{value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attributes}{name or names of numeric attributes to form a color legend. Multiples will be plotted in the same layer beside each other with a gap.}

\item{size}{the space (in lines) for this layer in the margin.}

\item{gap}{the space (in lines) between this layer and the prior layer.}

\item{size_p}{the proportion of the total axis which the legend should occupy (1 equates to the whole side)}

\item{gap_p}{the proportion of the total axis which would be spaced between multiple legends.}

\item{round}{the number of decimals included in the legend labels.}

\item{cex}{legend label size.}

\item{cex.title}{legend title size.}

\item{main}{legend title.}
}
\description{
Makes a simple color gradient legend corresponding to the main heatmap.
}
