% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_group_names}
\alias{lp_group_names}
\title{Plot group name layer}
\usage{
lp_group_names(
  lp,
  side,
  attribute,
  size = 1,
  gap = 0,
  col = "black",
  font = 1,
  adj = NULL,
  family = "",
  srt = NULL,
  cex = 0.8,
  show_bounding_box = F,
  plot_label = F
)
}
\arguments{
\item{lp}{layermap object.}

\item{side}{value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{name for the attribute which will be plotted in the layer. For group, this must be defined in column_groups or row_groups.}

\item{size}{the space (in lines) for this layer in the margin.}

\item{gap}{the space (in lines) between this layer and the prior layer.}

\item{col}{text color.}

\item{font}{text font.}

\item{adj}{text justification.}

\item{family}{text family.}

\item{srt}{text angle. Defaults as perpendicular to axis.}

\item{cex}{size value for names.}

\item{show_bounding_box}{shows the bounding space for this layer (useful for troubleshooting)}
}
\value{
layermap object
}
\description{
Function for plotting a name label for a group.
}
