% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{lp_group_names}
\alias{lp_group_names}
\title{Plot group name layer}
\usage{
lp_group_names(
  lp,
  side,
  attribute,
  col = "black",
  font = 1,
  adj = NULL,
  family = "",
  srt = NULL,
  size = 1,
  gap = 0,
  cex = 0.8,
  show_bounding_box = F,
  group_label = F
)
}
\arguments{
\item{lp}{- layermap object .}

\item{side}{- value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{col}{- named color vector, where the names are conditions found in the attribute.}

\item{srt}{}

\item{cex}{- cex value for label characters.}
}
\value{
layermap object
}
\description{
Function for plotting a name label for a group.
}
