% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_legend}
\alias{lp_legend}
\title{Plot simple categorical legend}
\usage{
lp_legend(
  lp,
  side,
  attribute,
  cex = 0.8,
  gap = 0.2,
  title = attribute,
  title.col = "black",
  title.cex = NULL,
  title.font = 3,
  ncol = "auto"
)
}
\arguments{
\item{lp}{layermap object}

\item{side}{value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{name for the attribute which is used for the legend. Defaults to "main", the main colors if categorical}

\item{cex}{total legend size.}

\item{title.col}{title color.}

\item{title.cex}{title size.}

\item{title.font}{title font.}

\item{ncol}{number of columns for the legend. Defaults to single columns for the sides and 3 for the top/bottom.}

\item{size}{the space (in lines) for this layer in the margin.}
}
\description{
Makes a simple color legend for a layermap object.
}
