% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_legend}
\alias{lp_legend}
\title{Plot simple categorical legend}
\usage{
lp_legend(
  lp,
  side,
  attributes = NULL,
  cex = 0.6,
  gap = 0.4,
  title.col = "black",
  title.cex = NULL,
  title.font = 3
)
}
\arguments{
\item{lp}{- layermap object}

\item{side}{- value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{- name for the attributes which will be plotted in the layer. Defaults to plotting all of them.}
}
\description{
Makes a simple color legend for a layermap object.
}
