% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layermap.R
\name{lp_names}
\alias{lp_names}
\title{Plot name layer}
\usage{
lp_names(
  lp,
  side,
  attribute = F,
  names = NULL,
  size = 1,
  gap = 0.4,
  autobox = T,
  cex = 0.8,
  show_bounding_box = F,
  just = "auto"
)
}
\arguments{
\item{lp}{- layermap object .}

\item{side}{- value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{- name for the attribute which will be plotted in the layer. Default is F, which plots the rownames.}
}
\value{
layermap object
}
\description{
Function for plotting an name layer based on column or row attributes.
}
