% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{lp_plot_values}
\alias{lp_plot_values}
\title{Plot values}
\usage{
lp_plot_values(lp, l_threshold = 50, round.n = 2, cex = 0.6)
}
\arguments{
\item{lp}{layermap object.}
}
\value{
layermap object
}
\description{
Function for plotting values over main heatmap
}
