% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap}
\alias{nheatmap}
\title{Plot nheatmap}
\usage{
nheatmap(
  value.df,
  xlim = NULL,
  ylim = NULL,
  column.df = NULL,
  row.df = NULL,
  column_groups = c(),
  row_groups = c(),
  palette = "PuOr",
  reverse_palette = T,
  zero_centered_colors = F,
  cluster_cols = F,
  cluster_rows = T,
  group_gap = 0.02,
  border = "grey25",
  plot_margin = c(0.2, 0.2, 0.2, 0.2),
  dim_reference = "din"
)
}
\arguments{
\item{value.df}{- numerical dataframe or matrix with column name and row names.}

\item{column.df}{- attribute dataframe which will be used for layer plotting functions on sides 1 and 3. Rownames correspond to value.df colnames.}

\item{row.df}{- attribute dataframe which will be used for layer plotting functions on sides 2 and 4. Rownames correspond to value.df rownames.}
}
\value{
nh object
}
\description{
Backbone function for nheatmap. This makes the basic heatmap structure and plots it. Requires a dataframe of just numerical values with named rows/columns. Row and column attributes are provided through additional dataframes keyed to the value dataframe names.
}
