% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap_annotate}
\alias{nheatmap_annotate}
\title{Plot annotation layer}
\usage{
nheatmap_annotate(
  nh,
  side,
  attribute,
  a.df = NULL,
  col = NULL,
  prop = 0.05,
  palette = "Viridis",
  show_bounding_box = F,
  type = "rect",
  label_just = "right",
  cex.label = 0.8
)
}
\arguments{
\item{nh}{- nheatmap object .}

\item{side}{- value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{- name for the attribute which will be plotted in the layer.}

\item{col}{- named color vector, where the names are conditions found in the attribute.}

\item{prop}{- the proportion of plotting space reserved for this layer}

\item{palette}{- hcl.colors palette to fill in unnamed conditions colors.}

\item{cex.label}{- cex value for label characters.}

\item{layer_just}{- the justification side for a layer label (right or left).}
}
\value{
nheatmap object
}
\description{
Function for plotting an annotation layer based on column or row attributes.
}
