% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nheatmap.R
\name{nheatmap_group}
\alias{nheatmap_group}
\title{Plot group layer}
\usage{
nheatmap_group(
  nh,
  side,
  attribute,
  col = NULL,
  palette = "Zissou 1",
  size = 1,
  gap = 0.4,
  cex = 0.8,
  show_bounding_box = F,
  label_just = "right",
  labels = T,
  cex.label = 0.8
)
}
\arguments{
\item{nh}{- nheatmap object .}

\item{side}{- value for which side of the plot to apply the layer (1-bottom, 2-left, 3-top, 4-right).}

\item{attribute}{- name for the attribute which will be plotted in the layer. For group, this must be defined in column_groups or row_groups.}

\item{col}{- named color vector, where the names are conditions found in the attribute.}

\item{palette}{- hcl.colors palette to fill in unnamed conditions colors.}

\item{labels}{- logical for whether group labels should be plotted.}

\item{prop}{- the proportion of plotting space reserved for this layer}

\item{layer_just}{- the justification side for a layer label (right or left).}

\item{cex_label}{- cex value for label characters.}
}
\value{
nheatmap object
}
\description{
Function for plotting a group layer based on column or row attributes.
}
